booking-usecase

+++ features/field/delivery/request.go
@@ -3,9 +3,9 @@ package delivery
 import "capstone-project/features/field"
 
 type FieldRequest struct {
-	VenueID  uint    `json:"venue_id" form:"venue_id"`
-	Category string  `json:"category" form:"category"`
-	Price    float64 `json:"price" form:"price"`
+	VenueID  uint   `json:"venue_id" form:"venue_id"`
+	Category string `json:"category" form:"category"`
+	Price    uint   `json:"price" form:"price"`
 }
 
 func ToCore(data FieldRequest) field.FieldCore {
